feat(dynamic-form): adds an optional wrapping element for the form item
you can now wrap an element with the its label and decsription in a containing element to help with
the structure and styling of the form

BREAKING CHANGE: additonalProps which used to me mapped to the form element are now mapped to the
wrapping element to keep the same API as the sections. To map additional props to the form element
use the formElementAdditionalProps attribute instead.