feat(dynamic-form): removes logic for conditional rendering in favour of a visibility utility
removes the limited hardcoded logic for conditional controlled by the rulerKey and condition
properties in favour of a setHidden utility which gives the user the power to control the visibiliy
in a more advanced and custom way

BREAKING CHANGE: removes some properties from the DynamicFormItem model. `value` renamed to
`initialValue`. Added `initiallyHidden` and removed `controlType: Hidden`. The return of
useDynamicForm is now an object. `setState` changed to `patchState` utility. `setErrors` and
`setWarnings` changed to `setIndicator`. Please check the commit for more detailed changes. Thank
you for using dynamicForm :)